Radically reworked the whole depth_first thing so that it properly does both forward and reverse checks.. I think now that opt_depot's recursive unification will now properly cohere with opt_clean's interpretation of when it's appropriate to collapse unification directories.
opt_depot still won't collapse pre-existing directories itself.. we'll
continue to rely on opt_clean to do that for us.

Also fixed a bug in depth_first that led it to miss unification
opportunities when the new package has a directory level that only
contains a directory which was not present under the old package.

w00t.

-- Jon